# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Ember local-first productivity API — core loop for CLI and MCP clients.
servers:
  - url: /api
    description: Base API path (proxied from frontend; default http://localhost:8080/api)
...
paths:
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by id
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project fields
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: archiveProject
      tags: [projects]
      summary: Archive project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /dragons/{id}/skills:
    get:
      operationId: listDragonSkills
      tags: [skills]
      summary: Skills available to a dragon with maturity state
      parameters:
        - $ref: "#/components/parameters/DragonId"
      responses:
        "200":
          description: Dragon skills with maturity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DragonSkillEntry"
        "404":
          $ref: "#/components/responses/NotFound"
  /dragons/{id}/skills/{skillId}/run:
    post:
      operationId: invokeDragonSkill
      tags: [skills]
      summary: Invoke a skill on behalf of a dragon
      parameters:
        - $ref: "#/components/parameters/DragonId"
        - $ref: "#/components/parameters/SkillId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeSkillRequest"
      responses:
        "201":
          description: Skill run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvokeSkillResponse"
        "402":
          description: Over monthly budget
        "403":
          description: Trust insufficient
        "404":
          $ref: "#/components/responses/NotFound"
        "428":
          description: High-cost confirmation required
        "503":
          description: AI unavailable
  /skill-runs/{id}:
    get:
      operationId: getSkillRun
      tags: [skills]
      summary: Get a skill run by id
      parameters:
        - $ref: "#/components/parameters/SkillRunId"
      responses:
        "200":
          description: Skill run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /skill-runs/{id}/verdict:
    post:
      operationId: recordSkillVerdict
      tags: [skills]
      summary: Keeper verdict on a skill run
      parameters:
        - $ref: "#/components/parameters/SkillRunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillVerdictRequest"
      responses:
        "200":
          description: Verdict recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillVerdictResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already verdicted
  /dragons/{id}/inbox:
    get:
      operationId: getDragonInbox
      tags: [skills]
      summary: Pending and recent skill runs for keeper review
      parameters:
        - $ref: "#/components/parameters/DragonId"
      responses:
        "200":
          description: Inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /dragons/{id}/skill-runs:
    get:
      operationId: listDragonSkillRuns
      tags: [skills]
      summary: Recent skill runs for a dragon
      parameters:
        - $ref: "#/components/parameters/DragonId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillRun"
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
    DragonId:
      name: id
      in: path
      required: true
      schema:
        type: string
    SkillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
    SkillRunId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DragonType:
      type: string
      enum: [cinder, moss, drift, frost]
    DragonStage:
      type: string
      enum: [egg, hatchling, adolescent, adult, ancient]
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        dragon_type: { $ref: "#/components/schemas/DragonType" }
        dragon_stage: { $ref: "#/components/schemas/DragonStage" }
        total_focus_minutes: { type: integer }
        project_summary: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_session_at: { type: string, format: date-time, nullable: true }
        last_decay_check: { type: string, format: date-time, nullable: true }
        is_archived: { type: integer }
      required:
        - id
        - name
        - dragon_type
        - dragon_stage
        - total_focus_minutes
        - project_summary
        - created_at
        - updated_at
        - is_archived
...
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string }
        project_summary: { type: string }
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    TrustBand:
      type: string
      enum: [paired, solo, autonomous]
    SkillMode:
      type: string
      enum: [paired, autonomous]
    SkillRunStatus:
      type: string
      enum: [pending, approved, edited, rejected, failed, over_budget]
    Verdict:
      type: string
      enum: [approve, edit, reject]
    Skill:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        agent_recipe_key: { type: string }
        default_trust_band: { $ref: "#/components/schemas/TrustBand" }
        cost_estimate_input_tokens: { type: integer }
        created_at: { type: string, format: date-time }
      required: [id, name, description, agent_recipe_key, default_trust_band, cost_estimate_input_tokens, created_at]
    DragonSkillMaturity:
      type: object
      properties:
        dragon_id: { type: string }
        skill_id: { type: string }
        runs: { type: integer }
        approvals: { type: integer }
        edits: { type: integer }
        rejections: { type: integer }
        current_trust: { $ref: "#/components/schemas/TrustBand" }
        locked_band: { $ref: "#/components/schemas/TrustBand", nullable: true }
        paused: { type: integer }
        last_used_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    DragonSkillEntry:
      type: object
      properties:
        skill: { $ref: "#/components/schemas/Skill" }
        maturity: { $ref: "#/components/schemas/DragonSkillMaturity" }
      required: [skill, maturity]
    SkillRun:
      type: object
      properties:
        id: { type: string }
        dragon_id: { type: string }
        skill_id: { type: string }
        project_id: { type: string }
        mode: { $ref: "#/components/schemas/SkillMode" }
        complexity: { type: string, enum: [simple, complex] }
        user_prompt: { type: string }
        system_prompt: { type: string }
        output_text: { type: string, nullable: true }
        user_edit: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/SkillRunStatus" }
        model: { type: string }
        input_tokens: { type: integer }
        output_tokens: { type: integer }
        cost_usd: { type: number }
        ran_at: { type: string, format: date-time }
        verdicted_at: { type: string, format: date-time, nullable: true }
      required:
        - id
        - dragon_id
        - skill_id
        - project_id
        - mode
        - user_prompt
        - status
        - ran_at
    InvokeSkillRequest:
      type: object
      properties:
        user_prompt: { type: string }
        mode: { $ref: "#/components/schemas/SkillMode" }
        confirm_high_cost: { type: boolean }
      required: [user_prompt]
    InvokeSkillResponse:
      type: object
      properties:
        ok: { type: boolean }
        run: { $ref: "#/components/schemas/SkillRun" }
        error: { type: string }
        required_trust: { $ref: "#/components/schemas/TrustBand" }
        current_trust: { $ref: "#/components/schemas/TrustBand" }
        estimated_cost_usd: { type: number }
    SkillVerdictRequest:
      type: object
      properties:
        verdict: { $ref: "#/components/schemas/Verdict" }
        status:
          type: string
          enum: [approved, edited, rejected]
        user_edit: { type: string }
      required: []
    SkillVerdictResponse:
      type: object
      properties:
        ok: { type: boolean }
        run: { $ref: "#/components/schemas/SkillRun" }
        error: { type: string }
    InboxItem:
      type: object
      properties:
        run: { $ref: "#/components/schemas/SkillRun" }
        skill_name: { type: string }
        pause_state: { type: boolean }
    InboxResponse:
      type: object
      properties:
        pending:
          type: array
          items: { $ref: "#/components/schemas/InboxItem" }
        recently_handled:
          type: array
          items: { $ref: "#/components/schemas/InboxItem" }
        paused_skills:
          type: array
          items:
            type: object
            properties:
              skill_id: { type: string }
              skill_name: { type: string }
        high_cost_pending:
          type: array
          items: { $ref: "#/components/schemas/InboxItem" }