# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ember local-first productivity API — core loop for CLI and MCP clients.
 * OpenAPI spec version: 0.2.0
 */
import type { TrustBand } from "./trustBand";

export interface Skill {
  id: string;
  name: string;
  description: string;
  agent_recipe_key: string;
  default_trust_band: TrustBand;
  cost_estimate_input_tokens: number;
  created_at: Date;
}