# lib/api-zod/src/generated/types/skillMode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ember local-first productivity API — core loop for CLI and MCP clients.
 * OpenAPI spec version: 0.2.0
 */

export type SkillMode = (typeof SkillMode)[keyof typeof SkillMode];

export const SkillMode = {
  paired: "paired",
  autonomous: "autonomous",
} as const;